Don't write js bundles into Jekyll source folder
They should be written directly to the build folder (i.e. `_site/`).

Without this commit, webpack writes to `src/assets/`, and these files are then copied to `_site/assets/` by Jeykll -- unless Jekyll gets there before webpack has finished, in which case it copies either no file (site breaks) or it copies an older file (very annoying for dev. work!).  Files must now be protected in the build folder (`_site/`) using the `keep_files` directive in `_config.yaml` so that Jeykll doesn't clean them out when building.  For the same reason, the order of webpack and Jeykll is reversed in the npm `build` script, so that webpack always runs after Jekyll is finished.