Fix `checkPropTypes` warning.
The first section of each play has an empty `startTime` key in its generated JSON file, which causes a `checkPropTypes` warning to be emitted, as `startTime` is marked `isRequired` (as below).  One solution would be to change the `parse.js` script to write in a value of 0.  Another solution is this one.

```
Warning: Failed prop type: The prop `startTime` is marked as required in `App`, but its value is `undefined`.
```