Implement <TimelineScrubber/> Component
This component should be visually identical to repurposed <TimelineIndicator/> component which was previously used in this position (with the exception of the cursor style, which I've changed to the more standard pointer, instead of `col-resize`).  It fixes a couple of issues (including #288 and #316), and is in general a more lightweight and maintainable implementation (no more dependency on `react-draggable`, either).