Persist score filters in localStorage
There are more sophisticated ways to implement this, but they would require more of a rewrite of the way redux is implemented in the codebase (in particular, the rootReducer would need to be refactored into a more best-practice-y style).  For our simple use case, this is adequate, and efficient -- there is also scope to persist, e.g., current video position timestamps in the same way.  Serializing the whole state with JSON.stringify is not the cheapest operation computationally (this best practice would be to use thunks and redux middleware -- but see above), so it's throttled to at most once every 2 seconds, which should be more than adequate.