Convert `play.jsx` from functional to class-based component
This will be necessary for what must follow.  The correct way to do this would be to have a modern functional component with hooks which is connected to the redux store and then to have the initial state assembled when the store is created, rather than having much of it passed it a component-creation time.  However, in the interest of not attempting to rewrite the entire application at this stage, this will afford an acceptable solution.