More flexible `utils.convertTimeToSeconds` (with modified tests)
The motivation here was to allow this function to handle `mm:ss` style timestamps.  The function is also now capable of tolerating `ss` formats and milliseconds, too.

Previously, timestamps that were in any way malformed would return `0`.  This implementation will return `0` only if it can parse nothing at all of the timestamp supplied.  The additional tolerance will result in incorrect time indices if mangled timestamps are passed, and I don't want to do anything about this for two reasons:
1) returning `0` is equally "incorrect";
2) implementing a check for perfectly-formed timestamps introduces an overhead that is unnecessary -- it should be implemented where such a check is actually necessary (see, e.g., `utils.validateTimestamp`).