Fix `font-family` value
Oddly using `sans` works on linux, but not on Mac...  In any event, `sans-serif` is the correct value.  There may be some mileage in using the same full font-stack used in the main site (basically "Source Sans Pro") but that would also mean hauling it in from Google Fonts, and seems a bit unnecessary for one small line of text.