Implement mousewheel events in TimelineScrubber
This commit implements the mousewheel events.  The increment or decrement is keyed to the length of the scrubber -- i.e., the duration of the shōdan.  Wheeling up or down advances or rewinds (retreats?!) the video by 2% of the section length.  This means that the (in-/de-)crement is longer for longer shōdan, and shorter for shorter shōdan.  I think this is reasonable.