Work-around for Chrome issue.
Fixes #462.

I'm pretty sure it's this one: https://bugs.chromium.org/p/chromium/issues/detail?id=927066

I hesitate to call it an outright bug, but it's certainly an implementation difference from all other browsers (it's probably a bug).

The fix/work-around is fine, and doesn't require any Javascript layout thrashing, which I had expected and feared would be the case.  In our case the issue was compounded by the fact that the `<Markup>` component we're using from the `interweave` package was rendering `<span>` elements (despite the documentation clearly stating that it defaults to `<div>` elements) which we were populating with block-level `<section>` elements and so on.