Instruct Istanbul to ignore untested `else` branch
I've spent too long trying to write a test to cover this `else` branch.  The problem is that the function is only called by an event handler attached to `document.body`, which is not handled by React's synthetic event system, and therefore cannot be simulated by the mish-mash of testing frameworks we're using in this codebase (ReactTestUtils, Enyzme, etc.).  I can fire the a "native" event (well, JSDom) to test the functionality, and that works (see the tests), but events simulated within React components by the testing frameworks do not bubble to the browser(/JSDom) level, and so it's impossible to trigger an event that satisfies the criteria for the uncovered `else` branch (i.e. that it comes from within the filters panel, and bubbles at the way to `document.body`).  There must be a way, of course, but I'm giving up now.