In Tabbed Narratives, replace text margins with padding
This ensures there is no gap between text blocks, and so the "flickering" of mouseover/mouseout behaviours is eliminated.  The fact that this also means that the on-hover background colour extends to include the padding is a pleasant side-effect.

One other side effect is slightly greater spacing between `<h2>` elements and immediately succeeding `<h3>` elements -- this occurs because previously the adjoining margins would collapse together, but adjacent padding does not exhibit this behaviour.  Of the available solutions, accepting the additional spacing seems best -- it's arguably a minor improvement.