Bump lots of dependencies
This is motivated by the "need" to address security vulnerabilities in `handlebars` and `serialize-javascript` ("need" is in quotes because `serialize-javascript` is only used by webpack, and `handlebars` is only used by Istanbul, so neither is in production).  Figured we might as well take as many safe minors as we can too, in preparation for launch.