Enable popoup video links in narrative sections
This is a hack because the Interweave library strips onclick attributes (despite no mention of this in the docs, and the use of `allowAttributes` and `disableFilters`, which the docs indicate should pass everything through, attributes beginning with "on" are still stripped).  There is presumably a solution available using a custom `Matcher` and rebuilding the `a[onclick]` element as a React component, but while the case could be made that that would be cleaner, it would add even more overhead to rendering these blocks, and we're already well down the rabbit-hole with this HTML munging, unfortunately.