Fix for missing Font Awesome character
Looks like the relevant icon has been removed, somehow?  We have a static mapping (of uncertain provenance) of codepoints to SASS variables in the codebase, so perhaps the version of the font files has been updated and the mapping hasn't?

This commit maps `$fa-var-external-link` (which doesn't exist in the version of "Font Awesome Free 5" we're linking to) to `\f35d`, the code point for " external-link-alt" (which does) -- see https://fontawesome.com/v5/cheatsheet.

Fixes #632.